separate code for definition of onvif type
