+++ xsd/onvif/onvif_analytics.go
@@ -1,0 +1,21 @@
+package onvif
+
+import (
+	"github.com/use-go/goonvif/xsd"
+)
+
+type AnalyticsDeviceCapabilities struct {
+	XAddr       xsd.AnyURI
+	RuleSupport xsd.Boolean
+	Extension   AnalyticsDeviceExtension
+}
+
+type AnalyticsDeviceExtension xsd.AnyType
+
+//AnalyticsCapabilities Check
+type AnalyticsCapabilities struct {
+	XAddr                  xsd.AnyURI
+	RuleSupport            xsd.Boolean
+	AnalyticsModuleSupport xsd.Boolean
+}
+type AnalyticsEngineConfigurationExtension xsd.AnyType
